src/main/java/org/projet/projetWeb/exceptions/GlobalExceptionController.java
package org.projet.projetWeb.exceptions;


import org.projet.projetWeb.dto.ErrorResponse;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@Slf4j
@ControllerAdvice
public class GlobalExceptionController {

    @ExceptionHandler(Exception.class)
    public ResponseEntity<ErrorResponse> exceptionHandler(Exception exception) {
        log.error("Une erreur s'est produite: {}", exception.getMessage());
        ErrorResponse errorResponse = new ErrorResponse("Erreur interne du serveur", exception.getMessage());
        return new ResponseEntity<>(errorResponse, HttpStatus.INTERNAL_SERVER_ERROR);
    }
}